refactor(service): split WebAuthnService into attestation + assertion ceremonies
The ~560-line WebAuthnService mixed the attestation (registration) and assertion
(authentication) ceremonies plus the shared WebAuthn library plumbing in one class.
Extract three focused collaborators behind the unchanged public surface:

- WebAuthnCeremonyFactory: shared serializer (cached) + ceremony-step manager factory
- AttestationService: registration options, attestation verification, credential storage
- AssertionService: assertion options (username-first / discoverable / decoy), discoverable
  user resolution, assertion verification

WebAuthnService becomes a thin facade delegating to the two services, so the controllers
and the authentication service keep their existing entry point and DI wiring is unchanged
(all three new services autowire). No behaviour change. Existing WebAuthnServiceTest now
builds the facade from the real collaborators (so the full delegation chain is exercised);
the private-method reflection tests target the owning service.

Verified: 577 unit tests pass, PHPStan level 10 + phpat architecture rules clean, CGL
clean, and the DI container compiles on a live TYPO3 v13 backend (login page 200).
