+++ Classes/Service/WebAuthnService.php
@@ -9,53 +9,29 @@ declare(strict_types=1);
 
 namespace Netresearch\NrPasskeysBe\Service;
 
-use Cose\Algorithm\Manager as AlgorithmManager;
-use Cose\Algorithm\Signature\ECDSA\ES256;
-use Cose\Algorithm\Signature\ECDSA\ES384;
-use Cose\Algorithm\Signature\ECDSA\ES512;
-use Cose\Algorithm\Signature\RSA\RS256;
 use Netresearch\NrPasskeysBe\Domain\Dto\AssertionOptions;
 use Netresearch\NrPasskeysBe\Domain\Dto\RegistrationOptions;
 use Netresearch\NrPasskeysBe\Domain\Dto\VerifiedAssertion;
 use Netresearch\NrPasskeysBe\Domain\Model\Credential;
-use Psr\Log\LoggerInterface;
 use RuntimeException;
-use Symfony\Component\Serializer\SerializerInterface;
-use Throwable;
-use Webauthn\AttestationStatement\AttestationStatementSupportManager;
-use Webauthn\AttestationStatement\NoneAttestationStatementSupport;
-use Webauthn\AuthenticatorAssertionResponseValidator;
-use Webauthn\AuthenticatorAttestationResponse;
-use Webauthn\AuthenticatorAttestationResponseValidator;
-use Webauthn\AuthenticatorSelectionCriteria;
-use Webauthn\CeremonyStep\CeremonyStepManagerFactory;
 use Webauthn\CredentialRecord;
-use Webauthn\Denormalizer\WebauthnSerializerFactory;
-use Webauthn\PublicKeyCredential;
 use Webauthn\PublicKeyCredentialCreationOptions;
-use Webauthn\PublicKeyCredentialDescriptor;
-use Webauthn\PublicKeyCredentialParameters;
 use Webauthn\PublicKeyCredentialRequestOptions;
-use Webauthn\PublicKeyCredentialRpEntity;
-use Webauthn\PublicKeyCredentialUserEntity;
-use Webauthn\TrustPath\EmptyTrustPath;
 
+/**
+ * Facade over the WebAuthn attestation (registration) and assertion
+ * (authentication) ceremonies.
+ *
+ * Keeps a single, stable entry point for the controllers and the authentication
+ * service while delegating the ceremony logic to the focused {@see AttestationService}
+ * and {@see AssertionService}. This preserves the public surface used across the
+ * extension; the ceremony detail lives in the dedicated services.
+ */
 final class WebAuthnService
 {
-    private const ALGORITHM_MAP = [
-        'ES256' => -7,
-        'ES384' => -35,
-        'ES512' => -36,
-        'RS256' => -257,
-    ];
-
-    private ?SerializerInterface $serializer = null;
-
     public function __construct(
-        private readonly ExtensionConfigurationService $configService,
-        private readonly ChallengeService $challengeService,
-        private readonly CredentialRepository $credentialRepository,
-        private readonly LoggerInterface $logger,
+        private readonly AttestationService $attestationService,
+        private readonly AssertionService $assertionService,
     ) {}
 
     /**
@@ -63,62 +39,11 @@ final class WebAuthnService
      */
     public function createRegistrationOptions(int $beUserUid, string $username, string $displayName): RegistrationOptions
     {
-        $rpId = $this->configService->getEffectiveRpId();
-        $rpName = $this->configService->getConfiguration()->getRpName();
-
-        $rp = PublicKeyCredentialRpEntity::create(
-            name: $rpName,
-            id: $rpId,
-        );
-
-        $userHandle = $this->createUserHandle($beUserUid);
-
-        $user = PublicKeyCredentialUserEntity::create(
-            name: $username,
-            id: $userHandle,
-            displayName: $displayName,
-        );
-
-        $challenge = $this->challengeService->generateChallenge();
-        $challengeToken = $this->challengeService->createChallengeToken($challenge);
-
-        $existingCredentials = $this->credentialRepository->findByBeUser($beUserUid);
-        $excludeCredentials = \array_map(
-            static fn(Credential $cred): PublicKeyCredentialDescriptor => PublicKeyCredentialDescriptor::create(
-                type: PublicKeyCredentialDescriptor::CREDENTIAL_TYPE_PUBLIC_KEY,
-                id: $cred->getCredentialId(),
-                transports: $cred->getTransportsArray(),
-            ),
-            $existingCredentials,
-        );
-
-        $authenticatorSelection = AuthenticatorSelectionCriteria::create(
-            userVerification: $this->configService->getConfiguration()->getUserVerification(),
-            residentKey: AuthenticatorSelectionCriteria::RESIDENT_KEY_REQUIREMENT_PREFERRED,
-        );
-
-        $options = PublicKeyCredentialCreationOptions::create(
-            rp: $rp,
-            user: $user,
-            challenge: $challenge,
-            pubKeyCredParams: $this->getPublicKeyCredentialParameters(),
-            authenticatorSelection: $authenticatorSelection,
-            attestation: PublicKeyCredentialCreationOptions::ATTESTATION_CONVEYANCE_PREFERENCE_NONE,
-            excludeCredentials: $excludeCredentials,
-            timeout: 60000,
-        );
-
-        return new RegistrationOptions(
-            options: $options,
-            challengeToken: $challengeToken,
-        );
+        return $this->attestationService->createRegistrationOptions($beUserUid, $username, $displayName);
     }
 
     /**
      * Verify a registration response from the browser.
-     *
-     * The $responseJson is the JSON-serialized PublicKeyCredential from the browser,
-     * already base64url-encoded as per the WebAuthn spec.
      *
      * @throws RuntimeException on verification failure
      */
@@ -129,71 +54,13 @@ final class WebAuthnService
         string $username,
         string $displayName,
     ): CredentialRecord {
-        $challenge = $this->challengeService->verifyChallengeToken($challengeToken);
-
-        $rpId = $this->configService->getEffectiveRpId();
-        $rpName = $this->configService->getConfiguration()->getRpName();
-        $userHandle = $this->createUserHandle($beUserUid);
-
-        $rp = PublicKeyCredentialRpEntity::create(name: $rpName, id: $rpId);
-        $user = PublicKeyCredentialUserEntity::create(
-            name: $username,
-            id: $userHandle,
-            displayName: $displayName,
-        );
-
-        $creationOptions = PublicKeyCredentialCreationOptions::create(
-            rp: $rp,
-            user: $user,
-            challenge: $challenge,
-            pubKeyCredParams: $this->getPublicKeyCredentialParameters(),
-            attestation: PublicKeyCredentialCreationOptions::ATTESTATION_CONVEYANCE_PREFERENCE_NONE,
-        );
-
-        // Deserialize the browser response
-        $publicKeyCredential = $this->getSerializer()->deserialize(
+        return $this->attestationService->verifyRegistrationResponse(
             $responseJson,
-            PublicKeyCredential::class,
-            'json',
+            $challengeToken,
+            $beUserUid,
+            $username,
+            $displayName,
         );
-
-        if (!$publicKeyCredential instanceof PublicKeyCredential) {
-            throw new RuntimeException('Failed to deserialize credential response', 1700000020);
-        }
-
-        $response = $publicKeyCredential->response;
-        if (!$response instanceof AuthenticatorAttestationResponse) {
-            throw new RuntimeException('Expected attestation response', 1700000021);
-        }
-
-        $factory = $this->createCeremonyFactory();
-        $ceremonyManager = $factory->creationCeremony();
-        $validator = AuthenticatorAttestationResponseValidator::create($ceremonyManager);
-
-        try {
-            $source = $validator->check(
-                authenticatorAttestationResponse: $response,
-                publicKeyCredentialCreationOptions: $creationOptions,
-                host: $rpId,
-            );
-        } catch (Throwable $e) {
-            $this->logger->error('Passkey registration verification failed', [
-                'be_user_uid' => $beUserUid,
-                'error' => $e->getMessage(),
-            ]);
-            throw new RuntimeException(
-                'Registration verification failed: ' . $e->getMessage(),
-                1700000022,
-                $e,
-            );
-        }
-
-        $this->logger->info('Passkey registered successfully', [
-            'be_user_uid' => $beUserUid,
-            'username' => $username,
-        ]);
-
-        return $source;
     }
 
     /**
@@ -201,32 +68,7 @@ final class WebAuthnService
      */
     public function createAssertionOptions(string $username, int $beUserUid): AssertionOptions
     {
-        $rpId = $this->configService->getEffectiveRpId();
-        $challenge = $this->challengeService->generateChallenge();
-        $challengeToken = $this->challengeService->createChallengeToken($challenge);
-
-        $credentials = $this->credentialRepository->findByBeUser($beUserUid);
-        $allowCredentials = \array_map(
-            static fn(Credential $cred): PublicKeyCredentialDescriptor => PublicKeyCredentialDescriptor::create(
-                type: PublicKeyCredentialDescriptor::CREDENTIAL_TYPE_PUBLIC_KEY,
-                id: $cred->getCredentialId(),
-                transports: $cred->getTransportsArray(),
-            ),
-            $credentials,
-        );
-
-        $options = PublicKeyCredentialRequestOptions::create(
-            challenge: $challenge,
-            rpId: $rpId,
-            allowCredentials: $allowCredentials,
-            userVerification: $this->configService->getConfiguration()->getUserVerification(),
-            timeout: 60000,
-        );
-
-        return new AssertionOptions(
-            options: $options,
-            challengeToken: $challengeToken,
-        );
+        return $this->assertionService->createAssertionOptions($username, $beUserUid);
     }
 
     /**
@@ -234,95 +76,23 @@ final class WebAuthnService
      */
     public function createDiscoverableAssertionOptions(): AssertionOptions
     {
-        $rpId = $this->configService->getEffectiveRpId();
-        $challenge = $this->challengeService->generateChallenge();
-        $challengeToken = $this->challengeService->createChallengeToken($challenge);
-
-        $options = PublicKeyCredentialRequestOptions::create(
-            challenge: $challenge,
-            rpId: $rpId,
-            allowCredentials: [],
-            userVerification: $this->configService->getConfiguration()->getUserVerification(),
-            timeout: 60000,
-        );
-
-        return new AssertionOptions(
-            options: $options,
-            challengeToken: $challengeToken,
-        );
+        return $this->assertionService->createDiscoverableAssertionOptions();
     }
 
     /**
-     * Create *decoy* assertion options for an unknown username.
-     *
-     * Returns options structurally identical to a real user's so the public
-     * login-options endpoint cannot be used to enumerate valid backend usernames.
-     * The decoy allowCredentials are derived deterministically from the username
-     * (HMAC over an HKDF-derived key from the extension encryption key), so repeated
-     * requests for the same unknown username yield stable, unguessable credential
-     * descriptors. A subsequent assertion against a decoy fails verification exactly
-     * as a wrong passkey would, keeping known and unknown users indistinguishable.
+     * Create *decoy* assertion options for an unknown username (user-enumeration defence).
      */
     public function createDecoyAssertionOptions(string $username): AssertionOptions
     {
-        $rpId = $this->configService->getEffectiveRpId();
-        $challenge = $this->challengeService->generateChallenge();
-        $challengeToken = $this->challengeService->createChallengeToken($challenge);
-
-        $key = $this->configService->getEncryptionKey();
-        $derivedKey = \hash_hkdf('sha256', $key, 32, 'nr_passkeys_be_decoy');
-        $decoyId = \hash_hmac('sha256', $username, $derivedKey, true);
-
-        $allowCredentials = [
-            PublicKeyCredentialDescriptor::create(
-                type: PublicKeyCredentialDescriptor::CREDENTIAL_TYPE_PUBLIC_KEY,
-                id: $decoyId,
-                transports: [],
-            ),
-        ];
-
-        $options = PublicKeyCredentialRequestOptions::create(
-            challenge: $challenge,
-            rpId: $rpId,
-            allowCredentials: $allowCredentials,
-            userVerification: $this->configService->getConfiguration()->getUserVerification(),
-            timeout: 60000,
-        );
-
-        return new AssertionOptions(
-            options: $options,
-            challengeToken: $challengeToken,
-        );
+        return $this->assertionService->createDecoyAssertionOptions($username);
     }
 
     /**
-     * Resolve the backend user UID from a passkey assertion response.
-     *
-     * Used for discoverable (usernameless) login where the credential ID
-     * in the assertion identifies the user without requiring a username.
+     * Resolve the backend user UID from a passkey assertion response (discoverable login).
      */
     public function findBeUserUidFromAssertion(string $responseJson): ?int
     {
-        try {
-            $publicKeyCredential = $this->getSerializer()->deserialize(
-                $responseJson,
-                PublicKeyCredential::class,
-                'json',
-            );
-
-            if (!$publicKeyCredential instanceof PublicKeyCredential) {
-                return null;
-            }
-
-            $credential = $this->credentialRepository->findByCredentialId($publicKeyCredential->rawId);
-            if ($credential === null || $credential->isRevoked()) {
-                return null;
-            }
-
-            return $credential->getBeUser();
-        } catch (Throwable) {
-            return null;
-        }
+        return $this->assertionService->findBeUserUidFromAssertion($responseJson);
     }
 
     /**
@@ -335,92 +105,7 @@ final class WebAuthnService
         string $challengeToken,
         int $beUserUid,
     ): VerifiedAssertion {
-        $challenge = $this->challengeService->verifyChallengeToken($challengeToken);
-        $rpId = $this->configService->getEffectiveRpId();
-
-        // Deserialize the browser response
-        $publicKeyCredential = $this->getSerializer()->deserialize(
-            $responseJson,
-            PublicKeyCredential::class,
-            'json',
-        );
-
-        if (!$publicKeyCredential instanceof PublicKeyCredential) {
-            throw new RuntimeException('Failed to deserialize assertion response', 1700000030);
-        }
-
-        $response = $publicKeyCredential->response;
-        if (!$response instanceof \Webauthn\AuthenticatorAssertionResponse) {
-            throw new RuntimeException('Expected assertion response', 1700000031);
-        }
-
-        // Find the credential by its ID
-        $credentialId = $publicKeyCredential->rawId;
-        $credential = $this->credentialRepository->findByCredentialId($credentialId);
-
-        if ($credential === null) {
-            $this->logger->warning('Assertion with unknown credential ID', [
-                'be_user_uid' => $beUserUid,
-            ]);
-            throw new RuntimeException('Unknown credential', 1700000032);
-        }
-
-        if ($credential->isRevoked()) {
-            throw new RuntimeException('Credential has been revoked', 1700000033);
-        }
-
-        if ($credential->getBeUser() !== $beUserUid) {
-            $this->logger->warning('Credential does not belong to the claimed user', [
-                'be_user_uid' => $beUserUid,
-                'credential_be_user' => $credential->getBeUser(),
-            ]);
-            throw new RuntimeException('Credential mismatch', 1700000034);
-        }
-
-        $storedSource = $this->credentialToSource($credential);
-
-        $requestOptions = PublicKeyCredentialRequestOptions::create(
-            challenge: $challenge,
-            rpId: $rpId,
-            userVerification: $this->configService->getConfiguration()->getUserVerification(),
-        );
-
-        $factory = $this->createCeremonyFactory();
-        $ceremonyManager = $factory->requestCeremony();
-        $validator = AuthenticatorAssertionResponseValidator::create($ceremonyManager);
-
-        try {
-            $updatedSource = $validator->check(
-                credentialRecord: $storedSource,
-                authenticatorAssertionResponse: $response,
-                publicKeyCredentialRequestOptions: $requestOptions,
-                host: $rpId,
-                userHandle: $credential->getUserHandle() !== '' ? $credential->getUserHandle() : null,
-            );
-        } catch (Throwable $e) {
-            $this->logger->error('Passkey assertion verification failed', [
-                'be_user_uid' => $beUserUid,
-                'error' => $e->getMessage(),
-            ]);
-            throw new RuntimeException(
-                'Assertion verification failed: ' . $e->getMessage(),
-                1700000035,
-                $e,
-            );
-        }
-
-        $this->credentialRepository->updateSignCount($credential->getUid(), $updatedSource->counter);
-        $this->credentialRepository->updateLastUsed($credential->getUid());
-
-        $this->logger->info('Passkey login successful', [
-            'be_user_uid' => $beUserUid,
-            'credential_uid' => $credential->getUid(),
-        ]);
-
-        return new VerifiedAssertion(
-            credential: $credential,
-            source: $updatedSource,
-        );
+        return $this->assertionService->verifyAssertionResponse($responseJson, $challengeToken, $beUserUid);
     }
 
     /**
@@ -431,21 +116,7 @@ final class WebAuthnService
         int $beUserUid,
         string $label,
     ): Credential {
-        $credential = new Credential(
-            beUser: $beUserUid,
-            credentialId: $source->publicKeyCredentialId,
-            publicKeyCose: $source->credentialPublicKey,
-            signCount: $source->counter,
-            userHandle: $source->userHandle,
-            aaguid: $source->aaguid->toString(),
-            transports: \json_encode($source->transports, JSON_THROW_ON_ERROR),
-            label: $label,
-        );
-
-        $uid = $this->credentialRepository->save($credential);
-        $credential->setUid($uid);
-
-        return $credential;
+        return $this->attestationService->storeCredential($source, $beUserUid, $label);
     }
 
     /**
@@ -453,7 +124,7 @@ final class WebAuthnService
      */
     public function serializeCreationOptions(PublicKeyCredentialCreationOptions $options): string
     {
-        return $this->getSerializer()->serialize($options, 'json');
+        return $this->attestationService->serializeCreationOptions($options);
     }
 
     /**
@@ -461,103 +132,6 @@ final class WebAuthnService
      */
     public function serializeRequestOptions(PublicKeyCredentialRequestOptions $options): string
     {
-        return $this->getSerializer()->serialize($options, 'json');
-    }
-
-    private function getSerializer(): SerializerInterface
-    {
-        if ($this->serializer === null) {
-            $attestationManager = $this->createAttestationStatementSupportManager();
-            $factory = new WebauthnSerializerFactory($attestationManager);
-            $this->serializer = $factory->create();
-        }
-
-        return $this->serializer;
-    }
-
-    private function createAttestationStatementSupportManager(): AttestationStatementSupportManager
-    {
-        $manager = new AttestationStatementSupportManager();
-        $manager->add(new NoneAttestationStatementSupport());
-
-        return $manager;
-    }
-
-    private function createCeremonyFactory(): CeremonyStepManagerFactory
-    {
-        $factory = new CeremonyStepManagerFactory();
-
-        $origin = $this->configService->getEffectiveOrigin();
-        $factory->setAllowedOrigins([$origin]);
-
-        $algorithmManager = $this->createAlgorithmManager();
-        $factory->setAlgorithmManager($algorithmManager);
-
-        $factory->setAttestationStatementSupportManager($this->createAttestationStatementSupportManager());
-
-        return $factory;
-    }
-
-    private function createAlgorithmManager(): AlgorithmManager
-    {
-        $algorithms = $this->configService->getConfiguration()->getAllowedAlgorithmsList();
-        $manager = AlgorithmManager::create();
-
-        foreach ($algorithms as $algo) {
-            match (\strtoupper(\trim($algo))) {
-                'ES256' => $manager->add(ES256::create()),
-                'ES384' => $manager->add(ES384::create()),
-                'ES512' => $manager->add(ES512::create()),
-                'RS256' => $manager->add(RS256::create()),
-                default => $this->logger->warning('Unknown algorithm configured', ['algorithm' => $algo]),
-            };
-        }
-
-        return $manager;
-    }
-
-    /**
-     * @return list<PublicKeyCredentialParameters>
-     */
-    private function getPublicKeyCredentialParameters(): array
-    {
-        $algorithms = $this->configService->getConfiguration()->getAllowedAlgorithmsList();
-        $params = [];
-
-        foreach ($algorithms as $algo) {
-            $algoId = self::ALGORITHM_MAP[\strtoupper(\trim($algo))] ?? null;
-            if ($algoId !== null) {
-                $params[] = PublicKeyCredentialParameters::createPk($algoId);
-            }
-        }
-
-        return $params;
-    }
-
-    private function createUserHandle(int $beUserUid): string
-    {
-        $key = $this->configService->getEncryptionKey();
-        $derivedKey = \hash_hkdf('sha256', $key, 32, 'nr_passkeys_be_user_handle');
-
-        return \hash_hmac('sha256', (string) $beUserUid, $derivedKey, true);
-    }
-
-    private function credentialToSource(Credential $credential): CredentialRecord
-    {
-        $aaguid = $credential->getAaguid() !== ''
-            ? \Symfony\Component\Uid\Uuid::fromString($credential->getAaguid())
-            : \Symfony\Component\Uid\Uuid::v4();
-
-        return CredentialRecord::create(
-            publicKeyCredentialId: $credential->getCredentialId(),
-            type: PublicKeyCredentialDescriptor::CREDENTIAL_TYPE_PUBLIC_KEY,
-            transports: $credential->getTransportsArray(),
-            attestationType: 'none',
-            trustPath: new EmptyTrustPath(),
-            aaguid: $aaguid,
-            credentialPublicKey: $credential->getPublicKeyCose(),
-            userHandle: $credential->getUserHandle(),
-            counter: $credential->getSignCount(),
-        );
+        return $this->assertionService->serializeRequestOptions($options);
     }
 }
